+++ docs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.21](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.20...@vtj/docs@0.19.21) (2026-08-27)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.19.20](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.19...@vtj/docs@0.19.20) (2026-08-27)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.19.19](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.17...@vtj/docs@0.19.19) (2026-08-27)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
 ## [0.19.18](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.17...@vtj/docs@0.19.18) (2026-08-26)
 
 **Note:** Version bump only for package @vtj/docs