+++ platforms/pro/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.21](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.19.20...@vtj/pro@0.19.21) (2026-08-27)
+
+
+### Features
+
+* **openapi:** 新增发布云端项目功能 ([abfde26](https://gitee.com/newgateway/vtj/commits/abfde261910f4c31800e574a8b61c00ddb5738bb))
+
+
+
+
+
+## [0.19.20](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.19.19...@vtj/pro@0.19.20) (2026-08-27)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.19.19](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.19.17...@vtj/pro@0.19.19) (2026-08-27)
+
+
+### Features
+
+* **vite:** 支持开发和预览服务的自定义 CORS 配置 ([36044df](https://gitee.com/newgateway/vtj/commits/36044df32e5035b35d8bbe4132af67f3ebf8deda))
+
+
+
+
+
 ## [0.19.18](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.19.17...@vtj/pro@0.19.18) (2026-08-26)